Defined new classes and rearrenged functions.
